requirements.txt
numpy<2.0
sqlalchemy == 2.0.36
pymysql == 1.1.1
paho-mqtt == 2.1.0
